refactor: Co-locate configurator-related files
