+++ app/components/chart-annotations-selector.tsx
@@ -1,6 +1,6 @@
 import { Box } from "@theme-ui/components";
 import React, { useEffect, useRef } from "react";
-import { ConfiguratorStateDescribingChart } from "../domain";
+import { ConfiguratorStateDescribingChart } from "../configurator";
 import { getFieldLabel } from "../domain/helpers";
 import { IconName } from "../icons";
 import { locales } from "../locales/locales";
