+++ app/components/chart-columns.tsx
@@ -1,6 +1,6 @@
 import React, { memo } from "react";
 import { Box } from "theme-ui";
-import { ColumnConfig, ColumnFields } from "../domain/config-types";
+import { ColumnConfig, ColumnFields } from "../configurator";
 import { Observation } from "../domain/data";
 import { isNumber } from "../domain/helpers";
 import {
@@ -10,7 +10,10 @@ import {
 import { useLocale } from "../lib/use-locale";
 import { A11yTable } from "./a11y-table";
 import { Tooltip } from "./charts-generic/interaction/tooltip";
-import { AxisWidthBand, AxisWidthBandDomain } from "./charts-generic/axis/axis-width-band";
+import {
+  AxisWidthBand,
+  AxisWidthBandDomain,
+} from "./charts-generic/axis/axis-width-band";
 import { AxisHeightLinear } from "./charts-generic/axis/axis-height-linear";
 import { ColumnsGrouped } from "./charts-generic/columns/columns-grouped";
 import { GroupedColumnChart } from "./charts-generic/columns/columns-grouped-state";
@@ -18,7 +21,10 @@ import { Columns } from "./charts-generic/columns/columns-simple";
 import { ColumnsStacked } from "./charts-generic/columns/columns-stacked";
 import { StackedColumnsChart } from "./charts-generic/columns/columns-stacked-state";
 import { ColumnChart } from "./charts-generic/columns/columns-state";
-import { ChartContainer, ChartSvg } from "./charts-generic/containers/containers";
+import {
+  ChartContainer,
+  ChartSvg,
+} from "./charts-generic/containers/containers";
 import { InteractionColumns } from "./charts-generic/overlay/overlay-columns";
 import { LegendColor } from "./charts-generic/legends/color";
 import { Loading, LoadingOverlay, NoDataHint } from "./hint";
