+++ app/components/chart-type-selector.tsx
@@ -5,7 +5,7 @@ import { getPossibleChartType } from "../domain";
 import {
   ChartType,
   ConfiguratorStateSelectingChartType,
-} from "../domain/config-types";
+} from "../configurator";
 import { useDataCubeMetadataWithComponentValuesQuery } from "../graphql/query-hooks";
 import { useLocale } from "../lib/use-locale";
 import { SectionTitle } from "./chart-controls/section";
