+++ app/components/charts-generic/bars/bars-grouped-state.tsx
@@ -9,11 +9,7 @@ import {
 } from "d3-scale";
 import * as React from "react";
 import { ReactNode, useCallback, useMemo } from "react";
-import {
-  BarFields,
-  SortingOrder,
-  SortingType,
-} from "../../../domain/config-types";
+import { BarFields, SortingOrder, SortingType } from "../../../configurator";
 import { Observation, ObservationValue } from "../../../domain/data";
 import { getPalette, mkNumber } from "../../../domain/helpers";
 import { sortByIndex } from "../../../lib/array";
