+++ app/components/charts-generic/use-chart-state.tsx
@@ -1,15 +1,16 @@
 import { createContext, useContext } from "react";
-import { ChartFields, Observation } from "../../domain";
+import { ChartFields } from "../../configurator";
+import { Observation } from "../../domain";
+import { ComponentFieldsFragment } from "../../graphql/query-hooks";
+import { AreasState } from "./areas/areas-state";
+import { GroupedBarsState } from "./bars/bars-grouped-state";
+import { BarsState } from "./bars/bars-state";
+import { GroupedColumnsState } from "./columns/columns-grouped-state";
+import { StackedColumnsState } from "./columns/columns-stacked-state";
 import { ColumnsState } from "./columns/columns-state";
 import { LinesState } from "./lines/lines-state";
-import { AreasState } from "./areas/areas-state";
-import { ScatterplotState } from "./scatterplot/scatterplot-state";
 import { PieState } from "./pie/pie-state";
-import { ComponentFieldsFragment } from "../../graphql/query-hooks";
-import { StackedColumnsState } from "./columns/columns-stacked-state";
-import { GroupedColumnsState } from "./columns/columns-grouped-state";
-import { BarsState } from "./bars/bars-state";
-import { GroupedBarsState } from "./bars/bars-grouped-state";
+import { ScatterplotState } from "./scatterplot/scatterplot-state";
 
 export interface ChartProps {
   data: Observation[];
