+++ app/components/data-download.tsx
@@ -3,7 +3,8 @@ import { Button } from "@theme-ui/components";
 import { csvFormat } from "d3-dsv";
 import { saveAs } from "file-saver";
 import React, { memo, ReactNode, useMemo } from "react";
-import { ChartConfig, ChartFields, Observation } from "../domain";
+import { ChartConfig, ChartFields } from "../configurator";
+import { Observation } from "../domain";
 import {
   ComponentFieldsFragment,
   useDataCubeObservationsQuery,
