+++ app/components/dataset-selector.tsx
@@ -3,7 +3,7 @@ import { I18n } from "@lingui/react";
 import { Box, Button, Flex, Text } from "@theme-ui/components";
 import React, { useRef, useState } from "react";
 import { useDebounce } from "use-debounce";
-import { useConfiguratorState } from "../domain";
+import { useConfiguratorState } from "../configurator";
 import { DataCubeResultOrder, useDataCubesQuery } from "../graphql/query-hooks";
 import { useLocale } from "../lib/use-locale";
 import { MiniSelect, SearchField } from "./form";
