+++ app/components/filters.tsx
@@ -1,10 +1,7 @@
 import { Trans } from "@lingui/macro";
 import { Box, Button } from "@theme-ui/components";
 import React, { useCallback } from "react";
-import {
-  getFilterValue,
-  useConfiguratorState,
-} from "../domain/configurator-state";
+import { getFilterValue, useConfiguratorState } from "../configurator";
 import { useDimensionValuesQuery } from "../graphql/query-hooks";
 import { useLocale } from "../lib/use-locale";
 import { MultiFilterField, SingleFilterField } from "./field";
