+++ app/docs/action-bar.docs.tsx
@@ -1,6 +1,6 @@
 import { markdown, ReactSpecimen } from "catalog";
 import { ActionBar } from "../components/action-bar";
-import { ConfiguratorStateProvider } from "../domain";
+import { ConfiguratorStateProvider } from "../configurator";
 import { states } from "./fixtures";
 import { Text, Box } from "@theme-ui/components";
