+++ app/domain/index.ts
@@ -1,6 +1,3 @@
-export * from "./configurator-state";
-export * from "./config-form";
-export * from "./config-types";
 export * from "./data";
 export * from "./charts";
 export * from "./chart-id";
