+++ app/pages/[locale]/__test/[chartId].tsx
@@ -1,6 +1,6 @@
 import { NextPage, GetStaticProps, GetStaticPaths } from "next";
 import { ChartPublished } from "../../../components/chart-published";
-import { Config } from "../../../domain/config-types";
+import { Config } from "../../../configurator";
 import { loadFixtureConfigIds, loadFixtureConfig } from "../../../test/utils";
 
 type PageProps = {
