+++ app/pages/[locale]/v/[chartId].tsx
@@ -11,7 +11,7 @@ import { ContentLayout } from "../../../components/layout";
 import { LocalizedLink } from "../../../components/links";
 import { PublishActions } from "../../../components/publish-actions";
 import { getConfig } from "../../../db/config";
-import { Config } from "../../../domain/config-types";
+import { Config } from "../../../configurator";
 import { useLocale } from "../../../lib/use-locale";
 
 type PageProps =
